requirements.txt
Django>=5.0
python-dotenv>=1.0
psycopg[binary]>=3.2
redis>=5.0
celery>=5.4
python-telegram-bot>=21.6
httpx>=0.27
Pillow>=10.4
rapidfuzz>=3.9
python-dateutil>=2.9
jinja2>=3.1
django-jazzmin>=3.0
itsdangerous>=2.2
openai>=1.43.0
dj-database-url>=2.2.0
